refactor(将post部分函数移动至单独的文件):
