+++ lib/utils/post.js
@@ -181,3 +181,14 @@ export async function processPostData(props, from) {
 
   delete props.allPages
 }
+
+/**
+ * 获取从第1页到指定页码的文章
+ * @param pageIndex 第几页
+ * @param list 所有文章
+ * @param pageSize 每页文章数量
+ * @returns {*}
+ */
+export const getListByPage = function(list, pageIndex, pageSize) {
+  return list.slice(0, pageIndex * pageSize)
+}
