+++ themes/heo/components/BlogPostListScroll.js
@@ -1,10 +1,10 @@
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { getListByPage } from '@/lib/utils'
 import { useEffect, useRef, useState } from 'react'
 import CONFIG from '../config'
 import BlogPostCard from './BlogPostCard'
 import BlogPostListEmpty from './BlogPostListEmpty'
+import { getListByPage } from '@/lib/utils/post'
 
 /**
  * 博客列表滚动分页
