+++ themes/matery/components/BlogPostListScroll.js
@@ -1,11 +1,11 @@
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { getListByPage } from '@/lib/utils'
 import throttle from 'lodash.throttle'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import CONFIG from '../config'
 import BlogPostCard from './BlogPostCard'
 import BlogPostListEmpty from './BlogPostListEmpty'
+import { getListByPage } from '@/lib/utils/post'
 
 /**
  * 博客列表滚动分页
